+++ nop-service-framework/nop-biz/src/main/java/io/nop/biz/BizConfigs.java
@@ -21,6 +21,10 @@ public interface BizConfigs {
     IConfigReference<Integer> CFG_BIZ_QUERY_IN_OP_MAX_ALLOW_VALUE_SIZE = varRef(
             s_loc, "nop.biz.query.in-op-max-allow-value-size", Integer.class, 100);
 
+    @Description("批量操作的入参集合最多允许的元素个数")
+    IConfigReference<Integer> CFG_BIZ_MAX_BATCH_SIZE = varRef(
+            s_loc, "nop.biz.max-batch-size", Integer.class, 500);
+
     @Description("最多允许前台发送几个左连接条件")
     IConfigReference<Integer> CFG_BIZ_QUERY_MAX_LEFT_JOIN_PROP_COUNT = varRef(
             s_loc, "nop.biz.query.max-left-join-prop-count", Integer.class, 3